backend/requirements.txt
python-dotenv
openai
fastapi
uvicorn
motor
passlib
bcrypt==4.1.2
python-jose
python-multipart